Squash NavListItem into NavMenuListItem, fixup styling
