+++ frontend/components/Header/index.tsx
@@ -1,15 +1,4 @@
 import Header from "./Header";
-import NavList from "./Nav/NavList";
-import NavListItem from "./Nav/NavListItem";
-import { NavMenuButton } from "./Nav/NavMenuButton";
-import NavMenuListItem from "./Nav/NavMenuListItem";
 import NavHeader from "./NavHeader";
 
-export {
-  Header,
-  NavHeader,
-  NavList,
-  NavMenuListItem,
-  NavListItem,
-  NavMenuButton,
-};
+export { Header, NavHeader };
